app/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-rest-swagger==2.2.0
django-cors-middleware==1.5.0
drf-nested-routers==0.92.5
drf-yasg==1.20.0
django-debug-toolbar==3.2
django-filter==2.4.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
sentry-sdk==0.19.4